Move articles to modals
