+++ client/src/components/utilities/Support.tsx
@@ -1,12 +1,12 @@
 import {ReactNode} from 'react';
-import {ArticleHeader} from '../components/layout/ArticleLayout';
+import {ArticleHeader} from '../layout/ArticleLayout';
 
 
 export default function Support() {
     return (
         <>
             <ArticleHeader>Student resources</ArticleHeader>
-            <section className="grid grid-cols-[repeat(auto-fill,_minmax(16rem,_1fr))]">
+            <section className="grid grid-cols-[repeat(auto-fill,_minmax(16rem,_1fr))] gap-2">
                 <SupportCard name="Crisis Text Line" tel="741-741" sms>
                     For everyone in crisis; text "Help", Text "LGBTQ" for LGBTQQ-specific support.
                 </SupportCard>
@@ -35,7 +35,7 @@ function SupportCard(props: SupportCardProps) {
     const {name, tel, children, sms} = props;
 
     return (
-        <div className="p-3">
+        <div className="px-5 py-3 bg-gray-100 dark:bg-content-secondary-dark rounded">
             <h2 className="font-semibold mb-0.5">{name}</h2>
             <h3><a href={`${sms ? 'sms' : 'tel'}:${tel}`}>{tel}</a></h3>
             <p className="secondary">{children}</p>
